SDK-1301: Reorganise files and add lint script
