+++ examples/doc-scan/package.json
@@ -6,6 +6,9 @@
   "engines": {
     "node": ">=10"
   },
+  "scripts": {
+    "lint": "node_modules/.bin/eslint *.js src/*/**.js"
+  },
   "dependencies": {
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
@@ -15,7 +18,7 @@
     "yoti": "file:../.."
   },
   "devDependencies": {
-    "eslint": "^4.19.1",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.18.0",
     "eslint-plugin-node": "^6.0.1"
