+++ examples/doc-scan/src/controllers/media.controller.js
@@ -1,4 +1,4 @@
-const config = require('../config');
+const config = require('../../config');
 
 const {
   DocScanClient,
@@ -7,14 +7,17 @@ const {
 module.exports = async (req, res) => {
   const docScanClient = new DocScanClient(
     config.YOTI_CLIENT_SDK_ID,
-    config.YOTI_PEM,
+    config.YOTI_PEM
   );
 
   try {
-    const media = await docScanClient.getMediaContent(req.session.DOC_SCAN_SESSION_ID, req.query.mediaId);
+    const media = await docScanClient.getMediaContent(
+      req.session.DOC_SCAN_SESSION_ID,
+      req.query.mediaId
+    );
     res.set('Content-Type', media.getMimeType());
     res.status(200).end(media.getContent().buffer);
-  } catch(error) {
+  } catch (error) {
     res.render('pages/error', { error });
   }
 };
